requirements.txt
# requirements.txt
python-dotenv
SQLAlchemy
PyMySQL